+++ app/Http/Controllers/PortofolioJualController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\SubscriberModel;
 use App\Models\User;
+use Gate;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\PortofolioJualModel;
@@ -46,7 +47,18 @@ class PortofolioJualController extends Controller
         return view('portofoliojual', $data);
     }
 
+    public function getDataAdmin()
+    {
+        $dataporto = PortofolioJualModel::join('tb_saham', 'tb_portofolio_jual.id_saham', '=', 'tb_saham.id_saham')
+            ->join('tb_sekuritas', 'tb_portofolio_jual.id_sekuritas', '=', 'tb_sekuritas.id_sekuritas')
+            ->join('users', 'tb_portofolio_jual.user_id', '=', 'users.id')
+            ->orderBy('user_id', 'asc')
+            ->get();
 
+        $data = compact(['dataporto']);
+        //dd($data);
+        return view('admin/portofoliojual', $data);
+    }
     public function getdataAnalyst($user_id)
     {
         $isSubscribed = SubscriberModel::where('id_subscriber', Auth::id())->where('id_analyst', $user_id)->where('status', 'subscribed')->first();
@@ -124,8 +136,11 @@ class PortofolioJualController extends Controller
         }
     }
 
-    public function getEdit($id_portofolio_jual)
+    public function getEdit($id_portofolio_jual, PortofolioJualModel $portoJual)
     {
+        if (!Gate::allows('update-delete-portojual', $portoJual)) {
+            abort(403);
+        }
         $dataporto = PortofolioJualModel::where('id_portofolio_jual', $id_portofolio_jual)
             ->join('tb_sekuritas', 'tb_portofolio_jual.id_sekuritas', '=', 'tb_sekuritas.id_sekuritas')
             ->join('tb_saham', 'tb_portofolio_jual.id_saham', '=', 'tb_saham.id_saham')->get();
@@ -138,9 +153,27 @@ class PortofolioJualController extends Controller
         return view('editportofoliojual', $data);
     }
 
-    public function editData(Request $request)
+    public function getEditAdmin($id_portofolio_jual)
     {
+        if (Auth::user()->id_roles == 1) {
+            $dataporto = PortofolioJualModel::where('id_portofolio_jual', $id_portofolio_jual)->join('tb_saham', 'tb_portofolio_jual.id_saham', '=', 'tb_saham.id_saham')->get();
+            $emiten = SahamModel::all();
+            $jenis_saham = JenisSahamModel::all();
+            $sekuritas = SekuritasModel::all();
 
+            $data = compact(['dataporto'], ['emiten'], ['jenis_saham'], ['sekuritas']);
+            //dd($data);
+            return view('admin/editportofoliojual', $data);
+        } else {
+            abort(403);
+        }
+    }
+
+    public function editData(Request $request, PortofolioJualModel $portoJual)
+    {
+        if (!Gate::allows('update-delete-portojual', $portoJual)) {
+            abort(403);
+        }
         $dataporto = PortofolioJualModel::where('id_portofolio_jual', $request->id_portofolio_jual)->first();
         //dd($dataporto);
 
@@ -155,10 +188,12 @@ class PortofolioJualController extends Controller
         $response = Http::acceptJson()
             ->withHeaders([
                 'X-API-KEY' => 'pCIjZsjxh8So9tFQksFPlyF6FbrM49'
-            ])->get('https://api.goapi.id/v1/stock/idx/' . $emiten)->json();
+            ])->get('https://api.goapi.id/v1/stock/idx/' . $emiten . '/historical', [
+                'to' => $request->tanggal_jual,
+                'from' => $request->tanggal_jual
+            ])->json();
 
-        $data = $response['data']['last_price'];
-        $closeprice = $response['data']['last_price']['close'];
+        $closeprice = $response['data']['results'][0]['close'];
         $harga_jual = $request->harga_jual;
         $close_persen = round((($harga_jual - $closeprice) / $harga_jual) * 100);
 
@@ -177,11 +212,66 @@ class PortofolioJualController extends Controller
 
     }
 
-    public function deleteData($id_portofolio_jual)
+    public function editDataAdmin(Request $request, PortofolioJualModel $portoJual)
     {
+        if (Auth::user()->id_roles == 1) {
+            $dataporto = PortofolioJualModel::where('id_portofolio_jual', $request->id_portofolio_jual)->first();
+            //dd($dataporto);
+
+            $id = Auth::id();
+            //dd($dataporto);
+            $getEmiten = SahamModel::select('nama_saham')
+                ->where('id_saham', $request->id_saham)
+                ->first();
+            $emiten = $getEmiten->nama_saham;
+
+            $response = Http::acceptJson()
+                ->withHeaders([
+                    'X-API-KEY' => 'pCIjZsjxh8So9tFQksFPlyF6FbrM49'
+                ])->get('https://api.goapi.id/v1/stock/idx/' . $emiten . '/historical', [
+                    'to' => $request->tanggal_jual,
+                    'from' => $request->tanggal_jual
+                ])->json();
+
+            //dd($response);
+            $closeprice = $response['data']['results'][0]['close'];
+            $harga_jual = $request->harga_jual;
+            $close_persen = round((($harga_jual - $closeprice) / $harga_jual) * 100);
+
+            $dataporto->id_saham = $request->id_saham;
+            $dataporto->jenis_saham = $request->id_jenis_saham;
+            $dataporto->volume = $request->volume;
+            $dataporto->tanggal_jual = $request->tanggal_jual;
+            $dataporto->harga_jual = $harga_jual;
+            $dataporto->id_sekuritas = $request->id_sekuritas;
+            $dataporto->close_persen = $close_persen;
+            $dataporto->save();
+
+
+            return redirect()->to('admin/portofoliojual/');
+        }
+        abort(403);
+
+    }
+
+    public function deleteData($id_portofolio_jual, PortofolioJualModel $portoJual)
+    {
+        if (!Gate::allows('update-delete-portojual', $portoJual)) {
+            abort(403);
+        }
         $dataporto = PortofolioJualModel::where('id_portofolio_jual', $id_portofolio_jual)->firstOrFail();
         $dataporto->delete();
         $id = Auth::id();
         return redirect()->to('portofoliojual/' . $id);
+    }
+
+    public function deleteDataAdmin($id_portofolio_jual)
+    {
+        if (Auth::user()->id_roles == 1) {
+            $dataporto = PortofolioJualModel::where('id_portofolio_jual', $id_portofolio_jual)->firstOrFail();
+            $dataporto->delete();
+            return redirect()->to('admin/portofoliojual');
+        }
+        abort(403);
     }
 }
